pkcs11-tool: port from the opensc suite
The PKCS#11 tool can be used to access compliant USB smart cards via the
PKCS#11 interface. The tool uses PCSC-lite as back end that in turn uses a USB
session back end. In order to access smart cards, the tool must be provided
with a dynamic PKCS#11 library through the command line arguments. The Sculpt
package exported by this GOA project uses the OpenSC PKCS#11 library that is
available as Genode port as well.